Move serializer classes to a more common place
These aren’t really specific to the platform settings, so putting them
at the settings module level makes sense.
